backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
requests
python-dotenv
streamlit
huggingface_hub
elevenlabs
moviepy
google-api-python-client 
google-auth-oauthlib
tweepy
itsdangerous
groq
pydub
google-cloud-texttospeech
google-auth
apscheduler
tavily
openai
python-multipart
sqlalchemy